apigw_route53/requirements.txt
aws-cdk-lib==2.34.2
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha